+++ sources/ippcp/pcpaes_xtsdecrypt_direct.c
@@ -174,6 +174,12 @@ IPPFUN(IppStatus, ippsAESDecryptXTS_Direct,(const Ipp8u* pSrc, Ipp8u* pDst, int 
          #if (_IPP>=_IPP_P8) || (_IPP32E>=_IPP32E_Y8)
          /* use Intel(R) AES New Instructions version if possible */
          if(AES_NI_ENABLED==RIJ_AESNI(&aesCtx)) {
+            #if(_IPP32E>=_IPP32E_K0)
+            if (IsFeatureEnabled(ippCPUID_AVX512VAES)) {
+               cpAESDecryptXTS_VAES(pDst, pSrc, encBlocks, RIJ_DKEYS(&aesCtx), RIJ_NR(&aesCtx), tweakCT);
+            }
+            else
+            #endif
             cpAESDecryptXTS_AES_NI(pDst, pSrc, encBlocks, RIJ_DKEYS(&aesCtx), RIJ_NR(&aesCtx), tweakCT);
             pSrc += encBlocks*AES_BLK_SIZE;
             pDst += encBlocks*AES_BLK_SIZE;